commit 4

+++ src/pages/header/header.js
@@ -2,17 +2,25 @@ import { useNavigate } from 'react-router-dom'
 import './header.css'
 const Header=(props)=>{
     const nav=useNavigate()
+    if(props.login===false){
+        return(<div className='fix-header'>
+            <div className="header">
+                <h3 className='watermark' onClick={()=>window.location.href='https://www.easterncoal.nic.in/'}>ECL</h3>
+                <h1 className='title'>Cyber-Security</h1>
+            </div>
+        </div>)
+    }
     return(<div className='fix-header'>
         <div className="header">
             <h3 className='watermark'>ECL</h3>
             <h1 className='title'>Cyber-Security</h1>
-            {(props.page==='dash' && props.login===true) && 
+            {props.page==='login' && 
                 <button className='header-button' onClick={()=>{
                     localStorage.clear()
-                    window.location.reload()
+                    nav('/')
                 }}>Logout</button>
             }
-            {(props.page==='quiz' || props.page==='result') && 
+            {props.page==='quiz' && 
             <button className="header-button" onClick={()=>nav('/dashboard')}>Go Back</button>}
             {props.page==='admin' && 
             <button className="header-button" onClick={()=>nav('/admin')}>Go Back</button>}
@@ -20,12 +28,8 @@ const Header=(props)=>{
             <button className="header-button" onClick={()=>nav('/question_list')}>Go Back</button>}
             {props.page==='user_res' && 
             <button className="header-button" onClick={()=>nav('/user_list')}>Go Back</button>}
-            {props.page==='admin_panel' && 
-            <button className='header-button' onClick={()=>{
-                localStorage.clear()
-                nav('/signin')
-            }}>Logout</button>
-            }
+            {props.page==='signin' &&
+            <button className="header-button" onClick={()=>nav('/')}>Go Back</button>}
         </div>
     </div>)
 }
